feat: npm cli - multi framework + Vercel AI SDK (#548)
* feat: initial commit

* fix: handleNextSelection call

* chore: format and lint

* chore: updated comments

* chore: changelog

* chore: updated readmes

* chore: added todo and removed MCP from Framework list for now

* chore: removed references to mcpServer and MCP. Will readd in its own PR

* chore: updated package-lock

* chore: removed missed mcpServer reference

* chore: lint
