+++ typescript/.changeset/slick-points-smile.md
@@ -1,0 +1,5 @@
+---
+"create-onchain-agent": patch
+---
+
+Added multi framework support
