+++ typescript/create-onchain-agent/README.md
@@ -24,6 +24,7 @@ This command will guide you through setting up an onchain agent project by promp
 ## Features
 
 - **Guided setup**: Interactive prompts help configure the project.
+- **Supports multiple AI frameworks**.
 - **Supports multiple blockchain networks**.
 - **Select your preferred wallet provider**.
 - **Preconfigured with Next.js, React, Tailwind CSS, and ESLint**.
@@ -34,9 +35,10 @@ The CLI will prompt you for the following details:
 
 1. **Project Name**: The name of your new onchain agent project.
 2. **Package Name**: The npm package name (auto-formatted if needed).
-3. **Network**: Choose from available blockchain networks.
-4. **Chain ID**: Specify if using a custom network.
-5. **Wallet Provider**: Select a preferred method for wallet management.
+3. **Framework**: Choose from available AI frameworks.
+4. **Network**: Choose from available blockchain networks.
+5. **Chain ID**: Specify if using a custom network.
+6. **Wallet Provider**: Select a preferred method for wallet management.
 
 After answering the prompts, the CLI will:
