+++ typescript/create-onchain-agent/templates/next/app/api/agent/agentkit/evm/cdp/prepare-agentkit.ts
@@ -1,0 +1,109 @@
+import {
+  AgentKit,
+  cdpApiActionProvider,
+  cdpWalletActionProvider,
+  CdpWalletProvider,
+  erc20ActionProvider,
+  pythActionProvider,
+  walletActionProvider,
+  WalletProvider,
+  wethActionProvider,
+} from "@coinbase/agentkit";
+import * as fs from "fs";
+
+/**
+ * AgentKit Integration Route
+ *
+ * This file is your gateway to integrating AgentKit with your product.
+ * It defines the core capabilities of your agent through WalletProvider
+ * and ActionProvider configuration.
+ *
+ * Key Components:
+ * 1. WalletProvider Setup:
+ *    - Configures the blockchain wallet integration
+ *    - Learn more: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit#evm-wallet-providers
+ *
+ * 2. ActionProviders Setup:
+ *    - Defines the specific actions your agent can perform
+ *    - Choose from built-in providers or create custom ones:
+ *      - Built-in: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit#action-providers
+ *      - Custom: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit#creating-an-action-provider
+ *
+ * # Next Steps:
+ * - Explore the AgentKit README: https://github.com/coinbase/agentkit
+ * - Experiment with different LLM configurations
+ * - Fine-tune agent parameters for your use case
+ *
+ * ## Want to contribute?
+ * Join us in shaping AgentKit! Check out the contribution guide:
+ * - https://github.com/coinbase/agentkit/blob/main/CONTRIBUTING.md
+ * - https://discord.gg/CDP
+ */
+
+// Configure a file to persist the agent's CDP MPC Wallet Data
+const WALLET_DATA_FILE = "wallet_data.txt";
+
+/**
+ * Prepares the AgentKit and WalletProvider.
+ *
+ * @function prepareAgentkitAndWalletProvider
+ * @returns {Promise<{ agentkit: AgentKit, walletProvider: WalletProvider }>} The initialized AI agent.
+ *
+ * @description Handles agent setup
+ *
+ * @throws {Error} If the agent initialization fails.
+ */
+export async function prepareAgentkitAndWalletProvider(): Promise<{
+  agentkit: AgentKit;
+  walletProvider: WalletProvider;
+}> {
+  try {
+    let walletDataStr: string | null = null;
+
+    // Read existing wallet data if available
+    if (fs.existsSync(WALLET_DATA_FILE)) {
+      try {
+        walletDataStr = fs.readFileSync(WALLET_DATA_FILE, "utf8");
+      } catch (error) {
+        console.error("Error reading wallet data:", error);
+        // Continue without wallet data
+      }
+    }
+
+    // Initialize WalletProvider: https://docs.cdp.coinbase.com/agentkit/docs/wallet-management
+    const walletProvider = await CdpWalletProvider.configureWithWallet({
+      apiKeyName: process.env.CDP_API_KEY_NAME,
+      apiKeyPrivateKey: process.env.CDP_API_KEY_PRIVATE_KEY,
+      networkId: process.env.NETWORK_ID || "base-sepolia",
+      cdpWalletData: walletDataStr || undefined,
+    });
+
+    // Initialize AgentKit: https://docs.cdp.coinbase.com/agentkit/docs/agent-actions
+    const agentkit = await AgentKit.from({
+      walletProvider,
+      actionProviders: [
+        wethActionProvider(),
+        pythActionProvider(),
+        walletActionProvider(),
+        erc20ActionProvider(),
+        cdpApiActionProvider({
+          apiKeyName: process.env.CDP_API_KEY_NAME,
+          apiKeyPrivateKey: process.env.CDP_API_KEY_PRIVATE_KEY,
+        }),
+        cdpWalletActionProvider({
+          apiKeyName: process.env.CDP_API_KEY_NAME,
+          apiKeyPrivateKey: process.env.CDP_API_KEY_PRIVATE_KEY,
+        }),
+      ],
+    });
+
+    // Save wallet data
+    const exportedWallet = await walletProvider.exportWallet();
+    fs.writeFileSync(WALLET_DATA_FILE, JSON.stringify(exportedWallet));
+
+    return { agentkit, walletProvider };
+  } catch (error) {
+    console.error("Error initializing agent:", error);
+    throw new Error("Failed to initialize agent");
+  }
+}
