+++ typescript/create-onchain-agent/templates/next/app/api/agent/agentkit/svm/solanaKeypair/prepare-agentkit.ts
@@ -1,0 +1,116 @@
+import {
+  ActionProvider,
+  AgentKit,
+  cdpApiActionProvider,
+  jupiterActionProvider,
+  SOLANA_NETWORK_ID,
+  SolanaKeypairWalletProvider,
+  splActionProvider,
+  walletActionProvider,
+  WalletProvider,
+} from "@coinbase/agentkit";
+import { Keypair } from "@solana/web3.js";
+import bs58 from "bs58";
+import fs from "fs";
+
+/**
+ * AgentKit Integration Route
+ *
+ * This file is your gateway to integrating AgentKit with your product.
+ * It defines the core capabilities of your agent through WalletProvider
+ * and ActionProvider configuration.
+ *
+ * Key Components:
+ * 1. WalletProvider Setup:
+ *    - Configures the blockchain wallet integration
+ *    - Learn more: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit#evm-wallet-providers
+ *
+ * 2. ActionProviders Setup:
+ *    - Defines the specific actions your agent can perform
+ *    - Choose from built-in providers or create custom ones:
+ *      - Built-in: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit#action-providers
+ *      - Custom: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit#creating-an-action-provider
+ *
+ * # Next Steps:
+ * - Explore the AgentKit README: https://github.com/coinbase/agentkit
+ * - Experiment with different LLM configurations
+ * - Fine-tune agent parameters for your use case
+ *
+ * ## Want to contribute?
+ * Join us in shaping AgentKit! Check out the contribution guide:
+ * - https://github.com/coinbase/agentkit/blob/main/CONTRIBUTING.md
+ * - https://discord.gg/CDP
+ */
+
+// Configure a file to persist a user's private key if none provided
+const WALLET_DATA_FILE = "wallet_data.txt";
+
+/**
+ * Prepares the AgentKit and WalletProvider.
+ *
+ * @function prepareAgentkitAndWalletProvider
+ * @returns {Promise<{ agentkit: AgentKit, walletProvider: WalletProvider }>} The initialized AI agent.
+ *
+ * @description Handles agent setup
+ *
+ * @throws {Error} If the agent initialization fails.
+ */
+export async function prepareAgentkitAndWalletProvider(): Promise<{
+  agentkit: AgentKit;
+  walletProvider: WalletProvider;
+}> {
+  try {
+    // Setup Private Key
+    let privateKey = process.env.SOLANA_PRIVATE_KEY as string;
+    if (!privateKey) {
+      if (fs.existsSync(WALLET_DATA_FILE)) {
+        privateKey = JSON.parse(fs.readFileSync(WALLET_DATA_FILE, "utf8")).privateKey;
+        console.info("Found private key in wallet_data.txt");
+      } else {
+        const keypair = Keypair.generate();
+        privateKey = bs58.encode(keypair.secretKey);
+        fs.writeFileSync(WALLET_DATA_FILE, JSON.stringify({ privateKey }));
+        console.log("Created new private key and saved to wallet_data.txt");
+        console.log(
+          "We recommend you save this private key to your .env file and delete wallet_data.txt afterwards.",
+        );
+      }
+    }
+
+    // Initialize WalletProvider: https://docs.cdp.coinbase.com/agentkit/docs/wallet-management
+    // Configure Solana Keypair Wallet Provider
+    const rpcUrl = process.env.SOLANA_RPC_URL;
+    let walletProvider: SolanaKeypairWalletProvider;
+    if (rpcUrl) {
+      walletProvider = await SolanaKeypairWalletProvider.fromRpcUrl(rpcUrl, privateKey);
+    } else {
+      const network = (process.env.NETWORK_ID ?? "solana-devnet") as SOLANA_NETWORK_ID;
+      walletProvider = await SolanaKeypairWalletProvider.fromNetwork(network, privateKey);
+    }
+
+    // Initialize AgentKit: https://docs.cdp.coinbase.com/agentkit/docs/agent-actions
+    const actionProviders: ActionProvider[] = [
+      walletActionProvider(),
+      splActionProvider(),
+      jupiterActionProvider(),
+    ];
+    const canUseCdpApi = process.env.CDP_API_KEY_NAME && process.env.CDP_API_KEY_PRIVATE_KEY;
+    if (canUseCdpApi) {
+      actionProviders.push(
+        cdpApiActionProvider({
+          apiKeyName: process.env.CDP_API_KEY_NAME,
+          apiKeyPrivateKey: process.env.CDP_API_KEY_PRIVATE_KEY,
+        }),
+      );
+    }
+    const agentkit = await AgentKit.from({
+      walletProvider,
+      actionProviders,
+    });
+
+    return { agentkit, walletProvider };
+  } catch (error) {
+    console.error("Error initializing agent:", error);
+    throw new Error("Failed to initialize agent");
+  }
+}
