+++ typescript/create-onchain-agent/templates/next/app/api/agent/framework/langchain/route.ts
@@ -1,0 +1,51 @@
+import { AgentRequest, AgentResponse } from "@/app/types/api";
+import { NextResponse } from "next/server";
+import { createAgent } from "./create-agent";
+/**
+ * Handles incoming POST requests to interact with the AgentKit-powered AI agent.
+ * This function processes user messages and streams responses from the agent.
+ *
+ * @function POST
+ * @param {Request & { json: () => Promise<AgentRequest> }} req - The incoming request object containing the user message.
+ * @returns {Promise<NextResponse<AgentResponse>>} JSON response containing the AI-generated reply or an error message.
+ *
+ * @description Sends a single message to the agent and returns the agents' final response.
+ *
+ * @example
+ * const response = await fetch("/api/agent", {
+ *     method: "POST",
+ *     headers: { "Content-Type": "application/json" },
+ *     body: JSON.stringify({ userMessage: input }),
+ * });
+ */
+export async function POST(
+  req: Request & { json: () => Promise<AgentRequest> },
+): Promise<NextResponse<AgentResponse>> {
+  try {
+    // 1️. Extract user message from the request body
+    const { userMessage } = await req.json();
+
+    // 2. Get the agent
+    const agent = await createAgent();
+
+    // 3.Start streaming the agent's response
+    const stream = await agent.stream(
+      { messages: [{ content: userMessage, role: "user" }] }, // The new message to send to the agent
+      { configurable: { thread_id: "AgentKit Discussion" } }, // Customizable thread ID for tracking conversations
+    );
+
+    // 4️. Process the streamed response chunks into a single message
+    let agentResponse = "";
+    for await (const chunk of stream) {
+      if ("agent" in chunk) {
+        agentResponse += chunk.agent.messages[0].content;
+      }
+    }
+
+    // 5️. Return the final response
+    return NextResponse.json({ response: agentResponse });
+  } catch (error) {
+    console.error("Error processing request:", error);
+    return NextResponse.json({ error: "Failed to process message" });
+  }
+}
