+++ typescript/create-onchain-agent/templates/next/app/api/agent/framework/vercel-ai-sdk/prepare-agentkit.ts
@@ -1,0 +1,19 @@
+import { AgentKit, WalletProvider } from "@coinbase/agentkit";
+
+/**
+ * A stub function that throws a "Not implemented" error.
+ *
+ * This is a placeholder function for the implementation that is selected.
+ * The actual implementation is selected by the cli user that generated this framework.
+ *
+ * @function prepareAgentkitAndWalletProvider
+ * @returns {Promise<{ agentkit: AgentKit, walletProvider: WalletProvider }>} The initialized AI agent.
+ *
+ * @throws {Error} Throws a "Not implemented" error.
+ */
+export async function prepareAgentkitAndWalletProvider(): Promise<{
+  agentkit: AgentKit;
+  walletProvider: WalletProvider;
+}> {
+  throw new Error("Not implemented");
+}
