+++ typescript/create-onchain-agent/templates/next/package.json
@@ -9,13 +9,16 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@ai-sdk/openai": "^1.2.1",
     "@coinbase/agentkit": "^0.3.0",
     "@coinbase/agentkit-langchain": "^0.3.0",
+    "@coinbase/agentkit-vercel-ai-sdk": "^0.1.0",
+    "@langchain/core": "^0.3.19",
     "@langchain/langgraph": "^0.2.21",
     "@langchain/openai": "^0.3.14",
-    "@langchain/core": "^0.3.19",
     "@solana/web3.js": "^1.98.0",
     "@tanstack/react-query": "^5",
+    "ai": "^4.1.54",
     "bs58": "^6.0.0",
     "next": "14.2.15",
     "react": "^18",
